Add JAX-based T-Digest implementation and refactor related modules (#90)
* Refactor `tdigest_normalizer` to improve modularity and performance by introducing feature-level processing and thread pool parallelization.


* Add JAX-based T-Digest Normalizer and corresponding unit tests


* Add JaxTDigest: A t-digest implementation leveraging JAX for fast, scalable quantile computation


* Refactor `TDigestModule`: replace host callbacks with pure JAX implementation.


* Update `TDigestNormalizer` tests to align with `TDigestModule` precision and initialization changes. Add new test cases for masking, batch shape handling, and inference mode.


* Remove unused `jax_tdigest_normalizer` and corresponding unit tests.


* Adjust merge condition in `jax_tdigest.py` to handle empty clusters


* Remove `fastdigest` dependency from `pyproject.toml`


* Adjust TDigestModule test configurations


* Update TDigestNormalizer to use Graph and HyperEdgeSet instead of JaxGraph and JaxHyperEdgeSet


* Replace `JaxGraph` and `JaxHyperEdgeSet` with `Graph` and `HyperEdgeSet` in test cases; update tests to include `JaxBackend`.


* Update normalizer and tdigest implementation to use float32 precision


* Apply stop_gradient to digest, xp_var, and slopes_var updates


* Add detailed docstrings to jax_tdigest.py functions and classes for clarity


* Add saturation handling to normalizers with configurable strategies


* Add unit tests for CenterReduceNormalizer and HyperEdgeSetCenterReduceNormalizer


* Refactor `test_simple_trainer` to improve readability and align dtype handling in cotangent test


* Add tests for `TDigestNormalizer` saturation modes (hard and soft)


* Refactor normalizers to simplify saturation logic and add update frequency support.


* Add validation tests for TDigest and CenterReduce normalizers


* Update `.coveragerc` omit rules to include new test file patterns


* Inline `JaxTDigest` into `tdigest_normalizer` and remove its standalone implementation.


* FIX Lint: Refactor normalization logic: extract update and saturation methods for reuse


---------

+++ tests/trainer/unit/test_simple_trainer.py
@@ -30,9 +30,11 @@ def create_tiny_model(context_structure):
     class SimpleDecoder(nnx.Module):
         def __call__(self, coordinates, graph, get_info=False):
             # No params here, just pass through
-            decision = Graph(backend=JaxBackend(),
+            decision = Graph(
+                backend=JaxBackend(),
                 hyper_edge_sets={
-                    "bus": HyperEdgeSet(backend=JaxBackend(),
+                    "bus": HyperEdgeSet(
+                        backend=JaxBackend(),
                         port_dict=None,
                         feature_array=coordinates,
                         feature_names={"phase_angle": jnp.array(0)},
@@ -64,7 +66,7 @@ def create_tiny_model(context_structure):
 
 def test_cast_cotangent_to_primal_dtype():
     primal = {"a": jnp.array([1.0], dtype=jnp.float32), "b": jnp.array([1], dtype=jnp.int32), "c": "not-an-array"}
-    cotangent = {"a": jnp.array([2.0], dtype=jnp.float64), "b": jnp.array([2.0], dtype=jnp.float32), "c": "not-an-array"}
+    cotangent = {"a": jnp.array([2.0], dtype=jnp.float32), "b": jnp.array([2.0], dtype=jnp.float32), "c": "not-an-array"}
 
     casted = _cast_cotangent_to_primal_dtype(cotangent, primal)
